refactor: move audio format definitions to backend.utilities.audio_formats
the enum is a pure leaf; living in _internal meant importing it pulled
the entire heavy _internal package __init__, which is what forced the
models ↔ _internal circular-import workaround. models/track.py now
imports AudioFormat at the top level like everything else.
